Move Uri and base marshalling from ServerRequestFactory to UriFactory.
